+++ src/main/java/ua/drozda/battlecity/core/TileUnit.java
@@ -8,7 +8,7 @@ import java.util.function.Function;
 public class TileUnit extends GameUnit {
     private TileType tileType;
 
-    public TileUnit(double x, double y, double width, double height, Long lives, Long currentTime, TileType tileType,
+    public TileUnit(double x, double y, double width, double height, Integer lives, Long currentTime, TileType tileType,
                     Function<GameUnit, Boolean> registerAction, Function<GameUnit, Boolean> unRegisterAction) {
         super(x, y, width, height, lives, BasicState.ACTIVE, registerAction, unRegisterAction, true);
         this.setTileType(tileType);
